PhoneNumber intden long a dönüştürüldü.

+++ ClassLibrary1/Core/Entities/Person.cs
@@ -7,10 +7,10 @@ namespace WorkoutControlling.Core.Entities
 		public int Id { get; set; }
 		public string FullName { get; set; }
 		public string Email { get; set; }
-		public int PhoneNumber { get; set; }
+		public long PhoneNumber { get; set; }
 		public DateTime RegistrationDate { get; set; } = DateTime.Now;
 
-		protected Person(int id, string fullName, string email, int phoneNumber)
+		protected Person(int id, string fullName, string email, long phoneNumber)
 		{
 			Id = id;
 			FullName = fullName;
